pyproject.toml
[project]
name = "python-fullstack-diploma"
version = "0.1.0"
description = ""
authors = [
    {name = "Donato4Humans",email = "donato.4work@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "djangochannelsrestframework (>=1.3.0,<2.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^7.0.0"

